+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2022-10-09
+
+### Added
+
+- Search strings starts with
+
+- Search strings ends with
+
 ## [0.4.1] - 2022-09-22
 
 - Added date filter: is on or after, ois on or before